requirements.txt
openpyxl==3.1.2
python-docx==1.1.0
weasyprint==57.2
jinja2==3.1.2
reportlab>=4.0.0
